Add withIdFn to specify ids for each element
In the event that a user wants to be able to set the id, they can use
withIdFn, where they provide a function that, given the deserialized
document, will allow them to return an id to be specified in the index
action. This allows the id to not be auto-generated, which means that we
can now update records.