Add new integration test for dynamic indexing
I'm currently making the assumption that the index will be created
automatically by the elasticsearch index if I try to send objects in the
bulk api into a new index. I didn't see anything to the contrary,
but that is an assumption I'm making that will need to be fixed if that
is not the case.